packet stats separate file
