+++ tests/p2p_test.cpp
@@ -91,11 +91,12 @@ TEST_F(P2PTest, capability_send_test) {
   network_conf.network_simulated_delay = 0;
   network_conf.network_bandwidth = 40;
   network_conf.network_transaction_interval = 1000;
-  auto thc1 = make_shared<TaraxaCapability>(host1, network_conf, GENESIS, false, addr_t(), nullptr, nullptr, nullptr,
-                                            nullptr, nullptr, nullptr, nullptr, nullptr, 2000);
+
+  auto thc1 = make_shared<TaraxaCapability>(host1, network_conf, GENESIS, addr_t(), nullptr, nullptr, nullptr, nullptr,
+                                            nullptr, nullptr, nullptr, nullptr, 2000);
   host1.registerCapability(thc1);
-  auto thc2 = make_shared<TaraxaCapability>(host2, network_conf, GENESIS, false, addr_t(), nullptr, nullptr, nullptr,
-                                            nullptr, nullptr, nullptr, nullptr, nullptr, 2000);
+  auto thc2 = make_shared<TaraxaCapability>(host2, network_conf, GENESIS, addr_t(), nullptr, nullptr, nullptr, nullptr,
+                                            nullptr, nullptr, nullptr, nullptr, 2000);
   host2.registerCapability(thc2);
   host1.start();
   host2.start();
@@ -153,11 +154,11 @@ TEST_F(P2PTest, capability_send_block) {
   network_conf.network_simulated_delay = 0;
   network_conf.network_bandwidth = 40;
   network_conf.network_transaction_interval = 1000;
-  auto thc1 = make_shared<TaraxaCapability>(host1, network_conf, GENESIS, false, addr_t(), nullptr, nullptr, nullptr,
-                                            nullptr, nullptr, nullptr, nullptr, nullptr, 2000);
+  auto thc1 = make_shared<TaraxaCapability>(host1, network_conf, GENESIS, addr_t(), nullptr, nullptr, nullptr, nullptr,
+                                            nullptr, nullptr, nullptr, nullptr, 2000);
   host1.registerCapability(thc1);
-  auto thc2 = make_shared<TaraxaCapability>(host2, network_conf, GENESIS, false, addr_t(), nullptr, nullptr, nullptr,
-                                            nullptr, nullptr, nullptr, nullptr, nullptr, 2000);
+  auto thc2 = make_shared<TaraxaCapability>(host2, network_conf, GENESIS, addr_t(), nullptr, nullptr, nullptr, nullptr,
+                                            nullptr, nullptr, nullptr, nullptr, 2000);
   host2.registerCapability(thc2);
   host1.start();
   host2.start();
@@ -231,14 +232,13 @@ TEST_F(P2PTest, block_propagate) {
   network_conf.network_bandwidth = 40;
   network_conf.network_transaction_interval = 1000;
 
-  auto thc1 = make_shared<TaraxaCapability>(host1, network_conf, GENESIS, false, addr_t(), nullptr, nullptr, nullptr,
-                                            nullptr, nullptr, nullptr, nullptr, nullptr, 2000);
+  auto thc1 = make_shared<TaraxaCapability>(host1, network_conf, GENESIS, addr_t(), nullptr, nullptr, nullptr, nullptr,
+                                            nullptr, nullptr, nullptr, nullptr, 2000);
   host1.registerCapability(thc1);
   std::vector<std::shared_ptr<TaraxaCapability>> vCapabilities;
   for (int i = 0; i < nodeCount; i++) {
-    vCapabilities.push_back(make_shared<TaraxaCapability>(*vHosts[i], network_conf, GENESIS, false, addr_t(), nullptr,
-                                                          nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, nullptr,
-                                                          2000));
+    vCapabilities.push_back(make_shared<TaraxaCapability>(*vHosts[i], network_conf, GENESIS, addr_t(), nullptr, nullptr,
+                                                          nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, 2000));
     vHosts[i]->registerCapability(vCapabilities[i]);
   }
   host1.start(true);
